feat: Product and Order Models and Controllers

+++ src/main/java/uptrip/repository/UserRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
 
     Boolean existsByUsername(String username);
@@ -19,5 +19,5 @@ public interface UserRepository extends JpaRepository<User, Integer> {
 
     List<User> findAllByRoles(Optional<Role> role);
 
-    User findUserById(int userId);
+    User findUserById(Long userId);
 }
